TDD-made WIP on HTTP Response Incremental Parsing
This initial implementation documents the development of an incremental
parsing implementation of HTTP Responses. This is a completely ground-up
implementation that is being developed in a test-driven manner. The aim
of this effort is two-fold:

  * Develop a functional incremental parser implementation for HTTP
    response messages parsing, to be used in the asynchronous HTTP
    client implementation.

  * Derive the minimal interface for incremental parsers, and provide a
    foundation on which to build concepts around incremental parsing for
    a potential library of these parsers.

To be truly generic, the approach aims to develop a specific
implementation of an incremental parser (in this case HTTP) and then see
if that can be applied to other protocols (i.e. XMPP, SMTP, etc.).