Enhancing Message Concept Documentation
This documentation update discusses the difference between directives,
modifiers, and wrappers. It also enhances the definition of the Message
Concept in preparation for the possible simplification of the
basic_request and basic_response implementations to turn them into POD
types similar to the POD types used by the HTTP server implementation.