Remove flush and close, change write semantics
With this commit we're changing the rules a little bit. The idea is a
write should immediately copy the range, linearize it in batches, and
schedule asynchronous writes in the process. In case there are any
errors encountered, this is reported to the handler primarily through a
mechanism like a function call to an explicit error handler in the
function.

In later revisions, it may be deemed that all calls to write should take
a completion handler function that takes an error code. This might make
the asynchronous handler programming a little more cumbersome, but will
open up doors to more powerful handler implementations.