Adding Overloads for Options on connection::write
This commit enables the user to call 'write' with a special callback
function. Instead of just passing in a range, the user can then provide
a callback function once the writes are done. This solution allows for
writing handlers that by themselves can chunk up responses to conserve
memory on their own.

An example streaming server using this approach should be doable.

This implementation is largely untested, but builds.

Also, setting the headers will by default send out the headers. In case
any errors are encountered at the time of the sending out of the
headers, that will be saved and then thrown in subsequent calls to
'write'.

This commit also contains a change to the macro that determines the
buffer size for connections, and for the chunk sizes of the linearized
buffers to reflect the correct type.