Completes (Untested) Asynchronous Connection
This commit actually completes the asynchronous connection
implementation. What remains to be done are:

- Complete the tests, make sure the server doesn't fail due to some
  serious asynchoronous handling issues.
- Write a static file server that uses HTTP/1.1 streaming

In this commit the whole system builds, but there needs to be a check on
whether the thing actually works. In this commit are:

  * A means for reading and scheduling an asynchronous read handler.
    Documentation is required on the API of the read handler.
  * A fully asynchronous event handling and thread-pool dispatching of
    user supplied handlers.

Expect a lot more changes starting this commit so hold your hats folks
it's going to be a bumpy ride.