Lorem Ipsum 10k Testing
This commit implements a real asynchronous HTTP/1.1 server. It still
doesn't support pipelined requests, but the connection framework is
there to be usable. Locally, this implementation has been benchmarked to
be able to handle 2800 requests per second, with Apache Bench (ab) with
the following parameters:

  ab -n 100000 -c 100 http://localhost:8000/

This was tested on a Linux machine, while the binary produced was built
with clang 2.9 (trunk).