Tag Table for HTTP Client Tags
This commit moves out the Table of Tags supported by the HTTP Client
``basic_client`` template. This table is shown in two locations, one in
the in-depth part of the HTTP implementation documentation, and another
at the HTTP Client API reference manual.

This commit also fleshes out the implementation section of the HTTP
Client API reference manual.