Implementing the Linearize Algorithm
This first part of the patch removes the linearization of the HTTP
Request algorithm from the bowels of the connection_helper detail that
exposes the linearize function to be used for all the client
implementations.

This refactoring is crucial to make the linerize algorithm a top-level
algorithm that deals with requests, responses, and normal messages. This
paves the way for a single point of extension for messages that have to
be linearized to a buffer.

Documentation and support for linearization of response objects will
follow.