Client or Server Tag Selector
This change introduces a selector metafunction that takes a tag, and
checks whether it is meant to be a client or server tag. The selector
metafunction then yields the actual tag found (whether it is a client or
server tag) so that it can be used for function tag dispatch purposes.