Removing Boost.Variant from Directives
This commit removes the reliance on Boost.Variant, and turns directives
into SFINAE aware directives based on the tag of the message. Basically,
what we want to happen is for code like the following to fail to
compile:

  basic_message<default_string> m;
  std::wstring w;
  m << source(w);

The reason it should fail to compile is because the default_string tag
tells basic_message to encapsulate std::string, not std::wstring objects
for the source. Previously, with the Boost.Variant solution this code
would compile and is therefore a serious bug.