Moving URI Parsers to Static Lib
This commit moves out the implementation of the URI parsing routines out
from the headers and into a single .cpp file. Now it means that
cpp-netlib is no longer just a header-only library. It also means that
this is proof enough that the compile times are really greatly affected
by the template metaprogramming happening under the hood of
Boost.Spirit.

Now things to do because of this commit:

* Overhaul the documentation regarding the installation and the
distribution of the library.

* Re-think the implementation of certain things like the static dispatch
of whether to use an asynchronous or synchronous client.

* Start making larger changes now that the compile-times are down when
building the tests.

That's it folks, header-only was fun while it lasted, unfortunately the
Boost.Spirit parts are taking too much a toll on the library's
usability. Hopefully this change makes the library more usable in a
larger context.