Moving Server-Specific Parsers to Static Lib
In line with the changes introduced to move the Boost.Spirit requiring
code out to static libraries, the parsers used by the HTTP server
implementation particularly the parsing of headers and the version have
been moved out into a static library.

This requires that all servers be linked against the static lib going
forward.