Making uri::port(...) return a convertible wrapper
To address the issue with regard to the Host header being sent by the
client to not include custom (or user-provided) port, the uri::port(...)
free function should instead return a convertible wrapper to either a
boost::optional<boost::uint16_t> or a boost::uint16_t. The rationale for
this is so that the client can eventually get for the
boost::optional<boost::uint16_t> instead of just a boost::uint16_t and
include that as part of the client's Host header forming routine.

That change will have to go into the linearize algorithm, and should be
done in the next commit. This change to the interface would still have
to be documented, but should otherwise not break any existing code
relying on uri::port(...) returning just a boost::uint16_t.