SFINAE for choosing 'write' overload
Aside from adding an additional clarification in the documentation for
the macro defining the size of a page, this change makes the user-facing
'write' implementation use SFINAE to choose whether to deal with the
buffer provided as a buffer that doesn't need to be copied (derives from
Boost.Asio's const_buffer type) or whether it's something that needs to
be copied (in the case of any other range).

In the future there should be other overloads which deal with explicit
Boost.Range iterator ranges, and maybe specific types.