Test for Less Copy Connection Write
This commit basically is a test for the case where the range passed to
the call to connection::write is actually a Boost.Asio const_buffer
derivative. This removes the requirement to make copies of the range
passed into the call to connection::write. This is a backward-compatible
change which means code that used to work with the range-based solution
will still be supported while Boost.Asio const_buffer derivatives passed
in will be supported as first-class elements that get used be the
implementation directly.

Note: Boost.Asio const_buffers explicitly do not own the memory that has
been wrapped to it. Care should be taken when passing in Boost.Asio
const_buffer sequences into the connection::write interface to make sure
that the data is still accessible until at least the provided callback
is provided.

Note: Backported from 0.9-devel into 0.8-devel.