Supporting BOOST_NETWORK_NO_LIB
This commit actually enables support for the BOOST_NETWORK_NO_LIB macro
that will allow users of the library to define this so that there will
be no need to have the externally-linked components when the network
library is used. This commit only adds the support for the URI parsing
parts -- this means that code using the HTTP client code don't need to
link the URI parser external library to use it.

NOTE: I'm still thinking about whether it's better that the default
implementation would remain header-only, and only have a special macro
for requiring the externally-linked libraries. I'm still not convinced
that the default for header-only would be productive, but I'm still
thinking about how to support some sort of backward compatibility for
the code that's already using the library in header only mode.