Making Server tests' port to run on configurable
This addresses some concerns with the unit tests for servers that don't
run mainly because the port they're all trying to run on (port 8000) is
being used across all the executables. Even if the socket address is
reused, that will only make sense for the same executable on UNIX/Linux.
Now the tests run on different ports, with the help of changes in the
tests and the CMake configuration.