SFINAE-ization of Tag Dispatch System
Thanks to Thomas Heller for the tips on the dramatic reduction of compile times in the cpp-netlib
implementation, this commit is now generally considered a major compile-time enhancement over
previous versions/iterations of the library. In this particular case the compile time reduction
is actually significant just by using SFINAE and partial specialization over (ab)use of the
Boost.MPL facilities.