Fixing cleanup order bug in pooled connection policy
Because of the refactorings that happened in the client implementation certain cleanup sequeneces
that used to not be an issue is now causing some errors especially with the pooled connection policy. These cleanup issues have been partly addressed in this commit, addressing a destruction bug that is manifested in the HTTP Client get tests when a synchronous HTTP Client with keepalive
support enabled in the tag used is destroyed.

The actual bug is that the io_service is destroyed first before the actual pooled connection policy
cleanup code is invoked, causing a segmentation fault with sockets bound to the io_service object
is destroyed because the internal state of the io_service referred to is already undefined by the
time the pool of sockets is destroyed.