Gutting the Beast Part 2.
In this commit I remove all traces of traits and metafunctions that can
all be replaced by dynamic dispatch. This also prepares to remove the
divergence between the synchronous and asynchronous implementation of
the messages/client. The goal is to actually just implement the
asynchronous version without bothering with a synchronous version. The
rationale for this is so that the implementation can grow and be generic
while at the same time require that the implementation rely on
Boost.Thread in a real way.

Part of this commit is the beginnings of the segmented request storage
base that will use Boost.IOStreams to support a unified interface to the
message implementation while making the underlying storage mechanism
replaceable and optimizable as releases go forward.

This all leads to simplification of the implementation while remaining
source-compatible as far as the interface is concerned.

Unfortunately all the interfaces will change, but hopefully majority of
the users will not have to change their usage of the library.