Message Refactoring
This commit gets the message tests to compile, but now the linker
complains that the implementation of the stubbed out interfaces for the
message implementation(s) are not available. This is a milestone which
indicates that the work required to get the library into working shape
would be clearer now.

More work lies ahead.