Major Surgery, aligning APIs.
So this set of changes moves forward the APIs of the internals of the
client implementation. This is going to be more in flux until the
request and response implementations get filled out. For the moment the
"simplification" continues.

WARNING: There's a lot of legacy code here still, most of which will not
be kept. After this refactoring process is completed, the layout of the
library will also change to better reflect the structure.