Step two: implement the resolver delegate factory.
This moves out the policies into the client/connection/ directory and
adds the necessary parts that deal with the creation of the
async_resolver implementation. We now have wired the async_resolver to
derive from a resolver_delegate type that the connections will use to
resolve IP addresses of hosts.