Getting the ducks in a neat row...
First, we get the asynchronous resolver delegate into a cpp file that
gets built and linked into the cppnetlib-client-connections lib.

We then  write it up in the connection factory and resolver delegate
factory implementation so that we can create instances of these
properly.

Then lastly we have the default implementation of the pure virtual
destructors for the pure interface types we use.