Mine Tripped: http::request and http::response.
The thing I've been delaying for a long time already is something I now
have to face. The good part about this is that there are concept
requirements that I can use to keep the interface and semantics from
changing. Unfortunately this also means that I need to come up with a
working http::request and http::response implementation that *does the
right thing* now that it's necessary.

The http::http_async_normal_connection is undergoing a lot of changes
especially now that the details of how the message stores the data has
been abstracted out of the client. It used to be that the wiring of the
promise objects and the future objects were evident in the
http_async_connection implementation and the http::response<async>
objects. That was bad design and now we're going all good OOP with this
implementation, which means we're going to have to hide these in the
http::response implementation.

This is yet another black hole I'm going to disappear in as I wire in an
efficient non-std::string storage mechanism for efficient small and
large payload processing in the http::response objects. The same goes
for supporting streams in the http::request objects.

Hang on tight, from here on out there will be a lot of changes that will
pertain to the http::request and http::response implementations that
may not be for the faint at heart.

Cheers