Updating the Concepts.
Due to the changes in the interfaces and removal for the requirement of
static tag-based dispatch, the concepts had to be updated to reflect the
now minimal requirements. This applies to the Message concept and the
ClientRequest and ServerRequest concepts.

For the same reason the modifiers for HTTP messages had to be updated to
reflect the interface. Same goes for the wrappers that deal specifically
with HTTP request objects.

The steps taken here will allow for more forward progress in making the
library more modular. Extension at a later time may require membership
in the hierarchy but this should make it more familiar using OOP
techniques. If it proves useful the wrappers, modifiers, and directives
can be written in a generic approach -- the meat though of the message
types would benefit from having a dynamic implementation for preserving
binary compatibility.