HTTP Connections Building.
After a few changes to the core body and headers wrappers in the message
implementation and the request linearization algorithm (to fix a bug
that makes the implementation not rely on the concept constraints) the
asynchronous HTTP connection implementation now compiles!

The bad news here is that the async_normal implementation still needs to
be cleaned up as much of the work here involved moving the
protocol_handler implementation into the actual connection
implementation. There's still a lot to do here to make the code more
readable which may mean just more documentation -- which I will be very
happy with as soon as things settle down.

This commit also limits the accessibility of the response field setting
interface that deals with futures and promises. Along with it is the fix
for the incremental parser's mis-handling of the edge case where the
range passed to the parse_until function is actually empty.