HTTP Client Connections now Build!
More changes in the message wrapper implementations but this is mostly
for consistency. This means though that it breaks backwards
compatibility for usage of these wrappers.

Ultimately this commit gets the asynchronous connection implementation
to a state where it actually builds and links. Next steps involve the
actual chore of making the request and result implementations work as
expected.