More Wrapper Implementations.
Here are a few more implementations of the wrappers required to get the
necessary bits of the implementation out of the way. This clears up some
fo the linker errors for the easier things that need to be dealt with.

We also introduce a new lib for the constants used throughout the
library. At some point later we're going to have to consolidate these
constants not only from the HTTP implementation but also the URI
implementation.