Progress: Client Constructor Test now Builds!
After some re-modeling of the internal implementation of the response
object as well as the interface used for initializing the promises, now
the client constructor test finally builds! This means most (not all) of
the request and response methods required just to instantiate an HTTP
client instance is ready for the linking.

Much more work to be done to:

1) Re-work and simplify the other tests.

2) Mock the delegates and remove reliance on actual external testing
that needs a network connection.

3) Make the request and response objects actually work. This is a tall
order just because of the sheer amount of code required to make this
happen.

4) Make the request and response objects more efficient in its use of
memory and synchronization primitives. I have doubts on the efficiency
and effectiveness of using futures for the data but this seems to be the
most straight-forward approach for the meantime.

Sounds like a lot of work to do but hopefully it's not as hard as it
reads.