Introducing client_options and request_options.
This commit makes changes to the API that moves the options setting to a
different type to keep the client API relatively stable even across
version releases supporting newer options. The intention is to keep the
ABI stable as well as we add more supported features in the form of
options in two levels: the client, and the request.

The client_options type is meant to define options that affect the
behavior of the constructed client and the connections it creates
internally. The request_options on the other hand define per-request
options pertaining to the HTTP implementation.

This paves the way for supporting more optional modes and extensions
that can and will be easily testable and turned on/off in a programmatic
manner.