Implement request_storage_base.
This commit implements an untested (but contains a building test)
request_storage_base that aims to hide the details of a segmented data
structure used to hold the request data. The intention is that the
storage base will be used if large amounts of data will be stored in an
efficient manner in memory.

This design also allows the request implementation to ignore the fact
that there's an underlying storage implementation and just use a
provided body chunk writer. This allows people to add a function that
pulls the data from whatever source and adds it to the request only when
necessary.

This is an evolutionary step forward moving from the inside out. The
intention is that we're able to build higher level logic to be able to
not rely on using the storage base in special circumstances.