Incremental Progress on HTTP Message
This time I've gotten all the tests to start building and linking. Now
it's the long road ahead to getting all the tests passing.

There's a number of API changes that need to be documented too. The
tests are now using opaque types and the virtual interfaces are clear
delineations for extension. At some point implementations of different
kinds of connection managers, connection delegates, connections
(websocket, spdy, proxied http, etc.), and other things would be nice to
have.

The next step probably is to get some sort of mocking around the
connections so that the HTTP parsing/handling logic can be handled
properly. Fake implementations of the connections now that they're
properly virtualized are going to be the other main focus going forward.