Long road to functional http::request.
One of the larger casualties of this refactor is http::request which
has been turned into a pimpl'ed value type. The implementation still has
a heavy-weight copy but at some point we can actually go ahead and do
some "copy-on-write" behavior if we really want to.

For now we're going to cover all the supported public APIs for the
http::request object. Once we're done with that we can then continue to
debugging, testing, and mocking out certain parts for the HTTP client
objects. I have a sneaking suspicion that if we get done with the
http::request and http::response objects then the HTTP clients might
"just work".