Additional cases for request_test.
This is just an incremental test case that uses a little of the mix
between the pimpl held in the request and the native storage that is
part of the request_storage_base from which the request derives from.
This is the default implementation and it's possible to implement a
different request type that uses different internal storage (probably a
better idea for mmap-io backed storage, etc.).