Fix bug with non-owned io_service deletion.
This was a bug waiting to happen that I've plugged. In case a non-owned
io_service was provided to the base, it will still delete it in the
constructor. I've changed it so that it only deletes the io_service in
case the base object owns the created io_service -- leaves it alone
otherwise.