Partial implementation of http::response.
After chasing down invalid memory access errors that shouldn't be there
I've made sure that we're turning on -Wall and -Werror for most (if not
all) the code in cpp-netlib built with CMAKE. If there's an easier way
to do this, it should be done. My CMake-foo is not strong enough to make
it happen except by doing copy-paste.

In other news this is the beginnings of the http::response type that's
supporting value semantics. There's still a number of stubbed out
functions regarding headers, which hopefully once implemented will allow
for moving forward to actually starting to debug the client
implementation.