More complete response functionality.
In this commit we add rudimentary header processing. The equality check
is not exactly correct as it's completely plausible that two different
response instances have different merged headers, but for the meantime
we're banking on internal equality rather than semantic equality.

What needs to happen to fix this is to add a little bit of
caching/merging logic to make sure that we're able to actually merge the
actual headers waiting in the future with the added headers manually
along with the removed headers, and compare those instead. Unfortunately
that's too much work for now and it's not really necessary at this time.