WIP: Tons of logging, improving build times.
I've reorganized a few things trying to hunt down why certain things
aren't happening as expected. I'm committing so that I can try and merge
from cpp-netlib/cpp-netlib:master to catch what Glyn has pushed with the
URI implementation. There's a few issues with trying to get the host and
other parts of the message through the wrappers and I'm trying a merge
to figure out if this has been fixed.