Lots of debugging.
At this point in the development, I'm stuck as to why the asynchronous
reads aren't returning data after writing the request down through the
sockets. I may just be missing some crucial setting on the created
sockets to properly enable the asynchronous processing, or the HTTP
request isn't well-formed, or there's some deadlock that I'm not seeing.
I'll debug this after a few hours of sleep.

Upped the version to 0.10.0a (a for the alpha) and hopefully this gets
functional sooner than later.