Lesson: Be Careful with Print Debugging
So it's been all good all along -- apparently in my eagerness to do
print debugging, actually getting the information from the stream_buf
actually causes the characters to be consumed. Removing the printing of
the linearized command_stream alleviates the hanging issue -- now I can
concentrate on other more pressing matters with the SSL implementation
and the other missing functionality with the response type.