Functional HTTP Client!
So, now this particular non-HTTPS implementation is doing what it's
supposed to be doing. There's still a bug in the HTTP implementation
which needs some squashing, but I'm proud to announce that we're now in
(almost) feature parity with the 0.9.3 implementation, *without the
header-only dogma*.

There's still quite a lot to do, but this is a milestone commit for the
refactoring effort.