Squashed Asio SSL FUBAR Issue
So, this affects the HTTPS implementation in 0.9.3 as well, where Asio
is not exactly being smart about how it deals with SSL errors. We've
had to do some acrobatics just to determine whether the error code
returned by an operation performed on an SSL stream resulted in a short
read. The change is simple, but tracking it down is so much of a PITA.

We probably should file a bug in Asio to be able to better detect the
situation with short reads on SSL streams, and somehow also file a bug
for it. At worst the short read should just use an asio::error::eof so
that it's not too hard to deal with this issue from the user side.

Also, this commit has some header-fudging to reduce dependencies and
incremental compile-times. A more extensive effort should be done to
reduce dependencies across header-files library-wide.