WIP: Asynchronous Server Implementation
This is an incremental installment into the asynchronous server
implementation revamp. The goal here is to isolate the parts that are
actually necessary to build client code and move more of the
implementation into externally-linked libraries. This allows us to
tweak, add features, and generally make the API/ABI more stable for
future extension and updates.