Asynchronous Server Builds!
So I set out to do one server test building, and I'm happy to say that
this commit does it. The other server tests just need to be jiggered a
bit to use the new interface.

This means we're a few steps closer to being able to merge to master!