WIP: Converting the Synchronous Server Implementation
Okay, I've hit a little road-block here. It seems that the synchronous
server implementation doesn't use the same parsing mechanism in the
async server implementation. This is going to have to change, but
before I go on I'd like to save the work I've already done to get to
this point.