examples are working with shared lib
'cppnetlib-utils-thread_pool' had to be added to the list of linked
libraries for 'cppnetlib-http-server'. For the static case this was not
a problem but in the shared case symbols were missing.

Listing external shared or static libraries in 'add_dependencies' is not
required. This CMake command is for adding dependencies between
different targets defined by 'add_executable', 'add_library' or
'add_custom_target'.