Various changes investigating short reads
This is in relation to issue #425 where a report of a short_read still
happening in Linux (Debian) with GCC/Clang. This couldn't be reproduced
with an OpenSSL acquired via Homebrew on Mac OS X.

More tests would require specific OpenSSL versions, which may or may
not exhibit this issue.