Doxygen+Sphinx+Breathe = Inline Documentation
This is the initial attempt at getting inline documentation in the
project implementation into the sphinx-generated documentation. The way
we achieve this is by doing the following:

  - Add michaeljones/breathe as a submodule, making it simpler to
    generate the documentation with just Sphinx installed on the
    developer machine.
  - Define a Doxyfile configuration which generates HTML and XML
    documentation. The generated XML files are needed by Breathe to
    generate the HTML as part of the Sphinx docs generation process.
  - Update the documentation for the client type. This is a starting
    point showing how the generated docs are in-lined in the RST docs
    via breathe.
  - Wire up the documentation generation in the Sphinx configuration.

To generate the documentation now, we need two things installed on the
machine generating the docs: Sphinx and Doxygen. The steps involved are:

  1. cd cpp-netlib && doxygen
  2. cd libs/network/doc/ && make html

The results should be in libs/network/doc/_build/html, and the
generated HTML for the HTTP Client reference page should include the
doxygen-generated docs.