Fix exception propaation for client errors (#754)
We change the cross-thread exception propagation given the asynchronous
client implementation. The previous implementation caused
double-wrapping of exceptions which works with the
std::shared_future<...> implementation, but not with
boost::shared_future<...> -- since Boost.Thread still uses
Boost.Exception for the APIs.

While we're here we make the exception propagation more explicit using
boost::rethrow_exception(...), so we can unwrap the exception that's
transported via an exception_ptr.

Fixes #749.